requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
httptools==0.6.4
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
llama-index-embeddings-openai
httpx~=0.28.1
tenacity~=9.0.0
openai~=1.61.1
llama-index-core~=0.12.16.post1